Sort functionality moved to ArticlesForm component
